massive update

+++ mods/revomon/pvp_keyword.py
@@ -1,3 +1,4 @@
+from typing import Any
 import asyncio
 import datetime
 import random
@@ -9,15 +10,19 @@ from discord.ext import commands
 from PIL import Image, ImageDraw, ImageFont
 
 from utils.helpers import respond
+from discord.embeds import Embed
+from discord.ext.commands.bot import Bot
+from typing import Dict, List, Optional, Union
+from unittest.mock import MagicMock
 
 
 class PvpLeaderboard(commands.Cog):
     def __init__(self, gradex: commands.Bot) -> None:
         self.gradex = gradex
-        self.rankings = {}
-        self.pvp_img = {}
+        self.rankings: Optional[List[Dict[str, Union[str, int]]]] = None
+        self.pvp_img: Dict[str, Any] = {}
 
-    def get_current_pvp_data(self):
+    def get_current_pvp_data(self) -> None:
         ranks = [
             "first",
             "second",
@@ -69,14 +74,14 @@ class PvpLeaderboard(commands.Cog):
 
         self.rankings = rankings_data
 
-    def update_pvp_image(self, data, output_path=None):
+    def update_pvp_image(self, data: Optional[List[Dict[str, Union[str, int]]]], output_path: None=None) -> None:
         img_width, cell_height = 1050, 50
         header_height = 40
         total_height = header_height + 15 * cell_height
         img = Image.new("RGB", (img_width, total_height), color=(0, 0, 0))
         draw = ImageDraw.Draw(img)
         try:
-            font = ImageFont.truetype("data/fonts/Cabal.ttf", 16)
+            font: Any = ImageFont.truetype("data/fonts/Cabal.ttf", 16)
         except Exception as e:
             print(f"Error loading font during update_pvp_image: {e}")
             font = ImageFont.load_default()
@@ -136,7 +141,7 @@ class PvpLeaderboard(commands.Cog):
         except Exception as e:
             print(f"Error converting PIL image during update_pvp_image: {e}")
 
-    def current_pvp_embed(self):
+    def current_pvp_embed(self) -> Embed:
         embed = discord.Embed(
             title=None,
             description=None,
@@ -147,13 +152,15 @@ class PvpLeaderboard(commands.Cog):
         embed.set_footer(text="Global Revomon Association")
         return embed
 
-    async def update_rankings(self):
+    async def update_rankings(self) -> None:
         try:
-            pvp_channel = await self.gradex.get_channel(1251368616662929529)
-            old_leaderboards = [message for message in pvp_channel.history(limit=2)]
+            pvp_channel = self.gradex.get_channel(1251368616662929529)
+            if pvp_channel is None or not hasattr(pvp_channel, "history"):
+                return
+            old_leaderboards = [message async for message in pvp_channel.history(limit=2)]
             for old_leaderboard in old_leaderboards:
                 await old_leaderboard.delete()
-            current_leaderboard = await pvp_channel.send(content="Loading...")
+            current_leaderboard = await pvp_channel.send(content="Loading...") # type: ignore
             while True:
                 try:
                     self.get_current_pvp_data()
@@ -180,19 +187,19 @@ class PvpLeaderboard(commands.Cog):
                     )
                 await asyncio.sleep(random.randint(600, 900))
         except Exception as e:
-            print(f"Error: {e}")
+            print(f"PvP Tracker Error: {e}")
 
     @commands.Cog.listener()
-    async def on_ready(self):
-        buttons = []
-        for button in buttons:
-            self.gradex.add_view(button)
+    async def on_ready(self) -> None:
+        # buttons = []
+        # for button in buttons:
+        #     self.gradex.add_view(button)
         print("Revomon Mod(PvP Leaderboard Tracker) is ready!")
         print("---------------------------")
         await self.update_rankings()
 
     @commands.Cog.listener()
-    async def on_message(self, message):
+    async def on_message(self, message: MagicMock) -> None:
         if message.author.bot:
             return
 
@@ -202,7 +209,7 @@ class PvpLeaderboard(commands.Cog):
                 self.get_current_pvp_data()
                 self.update_pvp_image(self.rankings)
                 embed = self.current_pvp_embed()
-                await respond(self.gradex, message=message, embed=embed, buttons=None)
+                await respond(self.gradex, message=message, embed=embed)
         except Exception as e:
             print(f"An error occurred during PvP Rankings Keyword on_message: {e}")
