+++ mods/the_elders_library/help_keyword.py
@@ -1,3 +1,4 @@
+from typing import Any
 import discord
 from discord.ext import commands
 
@@ -5,10 +6,10 @@ from utils.helpers import respond
 
 
 class help_keyword(commands.Cog):
-    def __init__(self, gradex):
+    def __init__(self, gradex: Any) -> None:
         self.gradex = gradex
 
-    def help_embed(self):
+    def help_embed(self) -> Any:
         embed = discord.Embed(
             title="Help Menu",
             description="*How can I help you?*",
@@ -21,7 +22,7 @@ class help_keyword(commands.Cog):
         return embed
 
     class gdex_help_buttons(discord.ui.View):
-        def __init__(self):
+        def __init__(self) -> None:
             super().__init__(timeout=None)
 
         @discord.ui.button(
@@ -30,8 +31,8 @@ class help_keyword(commands.Cog):
             custom_id="Mon Search Help",
         )
         async def mon_search_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             mon_search_help_embed = discord.Embed(
                 title="Revomon search",
                 description="""Type the name of the Revomon you would like more information about. Includes everything you"ll need to know, even Counterdex PvP Builds. Type **All Revomon** for a list of all the Revomon in my library.""",
@@ -52,8 +53,8 @@ class help_keyword(commands.Cog):
             custom_id="Move Search Help",
         )
         async def move_search_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             move_search_help_embed = discord.Embed(
                 title="Move search",
                 description="""Type the name of the move you would like more information about. Type **All Moves** for a list of all the moves in my library.""",
@@ -76,8 +77,8 @@ class help_keyword(commands.Cog):
             custom_id="PriMove Search Help",
         )
         async def primove_search_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             primove_search_help_embed = discord.Embed(
                 title="Priority Move search",
                 description="""Type the name of the Priority Move you would like more information about. Type **All Priority Moves** for a list of all the priority moves in my library.""",
@@ -100,8 +101,8 @@ class help_keyword(commands.Cog):
             custom_id="Ability Search Help",
         )
         async def ability_search_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             ability_search_help_embed = discord.Embed(
                 title="Ability search",
                 description="""Type the name of the Ability you would like more information about. Type **All Abilities** for a list of all the abilities in my library.""",
@@ -124,8 +125,8 @@ class help_keyword(commands.Cog):
             custom_id="Nature Search Help",
         )
         async def nature_search_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             nature_search_help_embed = discord.Embed(
                 title="Nature search",
                 description="""Type the name of the Nature you would like more information about. Type **All Natures** for a list of all the natures in my library.""",
@@ -148,8 +149,8 @@ class help_keyword(commands.Cog):
             custom_id="Item Search Help",
         )
         async def item_search_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             item_search_help_embed = discord.Embed(
                 title="Item search",
                 description="""Type the name of the item you would like more information about. Type **All Items** for a list of all the items in my library.""",
@@ -172,8 +173,8 @@ class help_keyword(commands.Cog):
             custom_id="Fruity Search Help",
         )
         async def fruity_search_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             fruity_search_help_embed = discord.Embed(
                 title="Fruity search",
                 description="""Type the name of the Fruity you would like more information about. Type **All Fruitys** for a list of all the fruitys in my library.""",
@@ -196,8 +197,8 @@ class help_keyword(commands.Cog):
             custom_id="Tier List Help",
         )
         async def tier_list_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             tier_list_help_embed = discord.Embed(
                 title="Tier List",
                 description="""Type **Tier List** to see the current meta tier list.""",
@@ -218,8 +219,8 @@ class help_keyword(commands.Cog):
             custom_id="EV Chart Help",
         )
         async def ev_chart_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             ev_chart_help_embed = discord.Embed(
                 title="EV Training Chart",
                 description="""Type **EV Chart** to find out ev_chartwhat EVs a Revomon gives once you defeat it.""",
@@ -240,8 +241,8 @@ class help_keyword(commands.Cog):
             custom_id="Spawn Locations Help",
         )
         async def spawn_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             spawn_help_embed = discord.Embed(
                 title="Spawn Locations",
                 description="""Type **Spawn** to see when & where all Revomon spawn.""",
@@ -262,8 +263,8 @@ class help_keyword(commands.Cog):
             custom_id="Pmon Help",
         )
         async def pmon_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             pmon_help_embed = discord.Embed(
                 title="Pokemon Counterparts",
                 description="""Type **Pokémon** for a list of Revomon and their Pokémon counterpart.""",
@@ -278,7 +279,7 @@ class help_keyword(commands.Cog):
             await interaction.followup.send(embed=pmon_help_embed, ephemeral=True)
 
     class help_buttons(discord.ui.View):
-        def __init__(self):
+        def __init__(self) -> None:
             super().__init__(timeout=None)
 
         @discord.ui.button(
@@ -287,8 +288,8 @@ class help_keyword(commands.Cog):
             custom_id="Gdex Help",
         )
         async def gdex_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             gdex_help_embed = discord.Embed(
                 title="Features & Instructions",
                 description=""">>> 1 | *The Gradex Tool combines the knowledge of The Elder & his vast library of books about the world of Revomon with the added convenience & power of Discord.*
@@ -324,8 +325,8 @@ class help_keyword(commands.Cog):
             custom_id="TTS Help",
         )
         async def tts_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             tts_help_embed = discord.Embed(
                 title="Tiptop's Top-up Shop",
                 description="""Type **Tiptop** for access to Tiptop's Top-up Shop. This is where you can purchase and/or redeem your Global Revomon NFTs like the Gradex Tool(Pro+), In-Game Currency(IGC) and more!!""",
@@ -349,8 +350,8 @@ class help_keyword(commands.Cog):
             custom_id="EED Help",
         )
         async def eed_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             eed_help_embed = discord.Embed(
                 title="Eazy's EV Daycare",
                 description="""Type **Eazy** for access to Eazy's EV Daycare. This is where you can purchase and/or redeem your EV Training NFTs to have your Revomon EV trained by Eazy.""",
@@ -374,8 +375,8 @@ class help_keyword(commands.Cog):
             custom_id="EA Help",
         )
         async def ea_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             ea_help_embed = discord.Embed(
                 title="Eleven's Arena",
                 description="""Type **Eleven** for access to Eleven's Arena. This is where you can find out more info about PvP and PvE events hosted by GRA.""",
@@ -397,8 +398,8 @@ class help_keyword(commands.Cog):
             custom_id="Titles Help",
         )
         async def titles_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             titles_help_embed = discord.Embed(
                 title="How to gain your NFT Titles",
                 description="""Follow the instructions in the {#collabland-join} channel to be awarded Titles that come with owning G.R.A. NFTs.""",
@@ -419,8 +420,8 @@ class help_keyword(commands.Cog):
             custom_id="OP Help",
         )
         async def op_help(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
             op_help_embed = discord.Embed(
                 title="GRA Online Presence",
                 description="""🔥 💯[OFFICIAL DAPP](https://www.globalrevomon.com/)
@@ -446,17 +447,18 @@ TheGlobalRevomonAssociation@Gmail.com""",
 
         @discord.ui.button(label="❌", style=discord.ButtonStyle.red, custom_id="exit")
         async def exit_embed(
-            self, interaction: discord.Interaction, Button: discord.ui.Button
-        ):
-            await interaction.message.delete()
+            self, interaction: discord.Interaction, Button: discord.ui.Button[Any]
+        ) -> None:
+            if interaction.message:
+                await interaction.message.delete()
 
     @commands.Cog.listener()
-    async def on_ready(self):
+    async def on_ready(self) -> None:
         print("The Elder's Library(Help Keyword) is ready!")
         print("---------------------------")
 
     @commands.Cog.listener()
-    async def on_message(self, message: discord.Message):
+    async def on_message(self, message: discord.Message) -> None:
         # Ignore messages from bots (including self)
         if message.author.bot:
             return
@@ -475,5 +477,5 @@ TheGlobalRevomonAssociation@Gmail.com""",
             print(f"An error occurred during on_message: {e}")
 
 
-async def setup(gradex: commands.Bot):
+async def setup(gradex: commands.Bot) -> None:
     await gradex.add_cog(help_keyword(gradex))
